Changed reconciliation types to double

+++ src/ui/modules/types/Balance.tsx
@@ -1,8 +1,8 @@
-import { int256 } from '@modules/types';
+import { double } from '@modules/types';
 
 export type Balance = {
   date: Date;
-  balance: int256;
+  balance: double;
   reconciled: boolean;
 };
 export type BalanceArray = Balance[];
